refs #36: 色変更ボタンの処理
