+++ package.json
@@ -27,6 +27,7 @@
     "react": "18.2.0",
     "react-chartjs-2": "^5.2.0",
     "react-dom": "18.2.0",
+    "react-input-color": "^4.0.1",
     "react-modal": "^3.16.1",
     "recoil": "^0.7.7",
     "typescript": "5.0.4"
