+++ src/components/ProjectShowUsersCard.tsx
@@ -1,8 +1,5 @@
 import { css } from '@emotion/react';
 import SystemUpdateAltIcon from '@mui/icons-material/SystemUpdateAlt';
-import DeleteForeverIcon from '@mui/icons-material/DeleteForever';
-import Image from 'next/image';
-import Link from 'next/link';
 import { useState } from 'react';
 import Modal from 'react-modal';
 import CloseIcon from '@mui/icons-material/Close';
@@ -12,10 +9,7 @@ import { useRouter } from 'next/router';
 import { useAuth } from '@/context/auth';
 import { createShowUserHandler } from '@/lib/firebase/createShowUserHandler';
 import { fetchShowUsersState } from '@/store/fetchShowUsersAtoms';
-import { githubUsers } from '@/store/atoms';
-import { githubUsersState } from '@/store/githubUsersAtom';
-import { deleteDoc, doc } from 'firebase/firestore';
-import { db } from '@/lib/firebaseClient';
+import ShowUser from './ProjectShowUsersCard/ShowUser';
 
 const styles = {
   container: css`
@@ -140,60 +134,6 @@ const styles = {
     border: 1px solid #ddd;
     font-weight: bold;
   `,
-  memberBody: css`
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    gap: 1rem;
-    padding: 0.5rem 1rem;
-    border: 1px solid #ddd;
-    border-top: none;
-    background-color: #f2f2f2;
-  `,
-  userInfo: css`
-    display: flex;
-    align-items: center;
-    gap: 1rem;
-    img {
-      border-radius: 50%;
-    }
-  `,
-  githubUserLink: css`
-    font-weight: bold;
-    font-size: 1.1rem;
-    color: #1f6feb;
-    transition: all 0.2s;
-    &:hover {
-      opacity: 0.8;
-    }
-  `,
-  githubUserName: css`
-    font-size: 0.9rem;
-    color: #7d8590;
-  `,
-  memberSettingButtonBox: css`
-    display: flex;
-    align-items: center;
-    gap: 1rem;
-  `,
-  changeColorButton: css`
-    width: 35px;
-    height: 35px;
-    border-radius: 50%;
-    border: 1px solid #ddd;
-    transition: all 0.2s;
-    &:hover {
-      opacity: 0.8;
-    }
-  `,
-  deleteButton: css`
-    width: 35px;
-    height: 35px;
-    transition: all 0.2s;
-    &:hover {
-      opacity: 0.8;
-    }
-  `,
 };
 
 const ProjectShowUsersCard = () => {
@@ -208,8 +148,6 @@ const ProjectShowUsersCard = () => {
   );
   const [fetchShowUsers, setFetchShowUsers] =
     useRecoilState(fetchShowUsersState);
-  const [githubUserList, setGitHubUserList] =
-    useRecoilState<githubUsers>(githubUsersState);
 
   const [githubId, setGithubId] = useState<string>('');
   const addMemberHandler = () => {
@@ -228,11 +166,6 @@ const ProjectShowUsersCard = () => {
   const closeAddMemberModal = () => {
     setAddMemberModalIsOpen(false);
     setGithubId('');
-  };
-
-  const deleteShowUser = async (showUserId: string) => {
-    const ref = doc(db, `showUsers/${showUserId}`);
-    await deleteDoc(ref);
   };
 
   return (
@@ -301,48 +234,10 @@ const ProjectShowUsersCard = () => {
           <span>Members</span>
         </div>
         {fetchShowUsers?.map((showUser) => (
-          <div css={styles.memberBody} key={showUser.id}>
-            <div css={styles.userInfo}>
-              <Image
-                src={
-                  githubUserList.find(
-                    (user) => user.id == showUser.githubId
-                  )?.avatarUrl ||
-                  '/images/defaultUserIcon.png'
-                }
-                alt={'github user icon'}
-                height={48}
-                width={48}
-              />
-              <div>
-                <Link
-                  href={`https://github.com/${showUser.githubId}`}
-                  css={styles.githubUserLink}
-                >
-                  {githubUserList.find(
-                    (user) => user.id == showUser.githubId
-                  )?.name || showUser.githubId}
-                </Link>
-                <p css={styles.githubUserName}>
-                  {showUser.githubId}
-                </p>
-              </div>
-            </div>
-            <div css={styles.memberSettingButtonBox}>
-              <button
-                css={[
-                  styles.changeColorButton,
-                  { backgroundColor: showUser.color },
-                ]}
-              ></button>
-              <button
-                css={styles.deleteButton}
-                onClick={() => deleteShowUser(showUser.id)}
-              >
-                <DeleteForeverIcon />
-              </button>
-            </div>
-          </div>
+          <ShowUser
+            key={showUser.id}
+            props={{ showUser }}
+          />
         ))}
       </div>
     </div>
